StreamLocalCopy: move to reader.h
Eventually move to internal header `internal/streamcopy.h`?

Based on 088e8173.
